Removed occlusion. Created a script attached to the prefab and containting the functions needed to update the mesh.
