+++ Assets/UnityARInterface/Prefabs/PlaneGridPrefab.prefab
@@ -37,6 +37,7 @@ GameObject:
   serializedVersion: 5
   m_Component:
   - component: {fileID: 4787822665528772}
+  - component: {fileID: 114531886388727584}
   m_Layer: 0
   m_Name: PlaneGridPrefab
   m_TagString: Untagged
@@ -126,3 +127,14 @@ MeshCollider:
   m_CookingOptions: 14
   m_SkinWidth: 0.01
   m_Mesh: {fileID: 0}
+--- !u!114 &114531886388727584
+MonoBehaviour:
+  m_ObjectHideFlags: 1
+  m_PrefabParentObject: {fileID: 0}
+  m_PrefabInternal: {fileID: 100100000}
+  m_GameObject: {fileID: 1372998203143648}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: b7bde3ae8ee6bf84eb583453d76b7dc9, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier:
