+++ Assets/UnityARInterface/Scripts/ARGrid.cs.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+guid: b7bde3ae8ee6bf84eb583453d76b7dc9
+timeCreated: 1516636273
+licenseType: Free
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
