+++ Assets/UnityARInterface/Scripts/ARPlaneVisualizerOccluding.cs.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+guid: bfc290aa0d7858745ac2eed2d7be108c
+timeCreated: 1505889486
+licenseType: Pro
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
